refactor to TS
